Added lots of new structure, start header writing
The heavy changes take place in InheritanceManager, which begins the job of
restructuring the Java AST into our own AST to handle inheritance.
HeaderWriter outputs defined_classes.h, which will be the header for all
classes used in the target java program. ImplementationWriter is empty, but
should eventually do the same for implementation files (They are the pretty
printers). PointChild.java is just a trivial inheritance test for when we get
Point.java working.
:

	new file:   HeaderWriter.java
	new file:   ImplementationWriter.java
	new file:   InheritanceManager.java
	modified:   Makefile
	modified:   QimppTranslator.java
	new file:   testfiles/PointChild.java
	renamed:    testfileout/main.cc -> testfiles/main.cc