Added InheritanceTreeManager (incomplete)
InheritanceTreeManager maintains a data structure with all defined
classes and their methods, so that we can get an ordered list of the
inheritance hierarchy for any object for InheritanceManager, and
all necessary methods and fields can be included. It is incomplete.

Also added: InheritanceTreeManagerTest and SimpleTest (a reference test
from junit)